Add files via upload
Two deep neural networks were created to detect the level of bias in reviews on the site Metacritic for the best and worst 25 video games of all time. One model was trained using Amazon reviews, the other model was trained using data from Metacritic which was collected using a web scraping algorithm was created to collect reviews from the site.